+++ lib/concierge/repositories/sync_process_repository.rb
@@ -6,6 +6,7 @@
 # in this table.
 class SyncProcessRepository
   include Hanami::Repository
+  extend  Concierge::Repositories::Pagination
 
   # returns the total number of properties stored in the database.
   def self.count
@@ -14,9 +15,21 @@ class SyncProcessRepository
 
   def self.recent_successful_sync_for_host(host)
     query do
-      where(successful: true).
-      where(host_id: host.id).
+      where(successful: true, host_id: host.id)
+    end.most_recent
+  end
+
+  # orders the collection by its start time, most recent first.
+  def self.most_recent
+    query do
       desc(:started_at)
+    end
+  end
+
+  # queries for synchronisation process of the given +type+.
+  def self.of_type(type)
+    query do
+      where(type: type)
     end
   end